Rename docs package
